Fixed major issue where RunningBoard would suddenly use another PID
This rendered our added properties useless, not being able to forward information about any immortal processes to SpringBoard when the latter loads.

The solution is to add properties to RBSProcessHandle (for RunningBoard) and RBSProcessState (for SpringBoard). This allows us to always have access to the information whether a process is immortal or if media is playing (mentioned as "partying").