Move init hooks out of RRManager
An anonymous c++ struct is used to store the different functions (normal, welcome, trial, pirated). By not naming the struct, the intentions of the struct is less clear when inspecting in runtime. To not make it easy to just replace the struct's other pointers with the first one, it isn't sent on the objc_msgSend boundary, but instead used as an extern global variable which will be harder to find and patch for script kiddies.